requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-crontab==0.7.1
django-currentuser==0.5.2
django-excel-response==2.0.5
django-filter==2.4.0
django-group-by==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mptt==0.11.0
django-online-users==0.3
django-scheduler==0.9.3
django-uuslug==1.2.0
et-xmlfile==1.0.1
icalendar==4.0.7
idna==2.10
jdcal==1.4.1
openpyxl==2.4.11
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
PyMySQL==1.0.2
python-dateutil==2.8.1
python-magic-bin==0.4.14
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
